+++ hortum_mobile/lib/views/edit/components/edit_picture.dart
